+++ src/main/java/eu/eutampieri/catacombs/ui/EndGameState.java
@@ -20,7 +20,7 @@ public class EndGameState extends State {
     }
 
     @Override
-    public void update (final float delta) {
+    public void update (final int delta) {
         this.blinkDelayCount *= delta;
         if (this.blinkDelayCount >= BLINK_DELAY) {
             this.blinkDelayCount = 0;
